Refactor reducer tests into seperate files
