+++ test/store/reducers/common.js
@@ -1,0 +1,18 @@
+export const mapId = 'map-id';
+export const layerId = 'layer-id';
+export const layerId2 = 'layer-id-2';
+export const layer = {
+  id: layerId,
+  visible: true,
+  nondefault: true,
+  loaded: false,
+  credit: 'credit',
+};
+export const layer2 = {
+  id: layerId2,
+  visible: true,
+  nondefault: false,
+  loaded: false,
+  credit: 'credit',
+};
+export const timeseries = { [layerId]: { adminFilter: false } };
